requirements.txt
torch~=2.2.1
py2neo~=2021.2.4
# icecream~=2.1.3
pyyaml~=6.0.1
python-dotenv~=1.0.1
mysql-connector-python==9.1.0
zhipuai~=2.0.1
loguru~=0.7.2
multipledispatch~=1.0.0
fastapi~=0.109.2
uvicorn~=0.27.1
gradio~=4.29.0
pyahocorasick~=2.0.0
apscheduler~=3.10.4
requests~=2.31.0
edge-tts~=6.1.10
langchain~=0.1.8
langchain-openai~=0.0.7
transformers~=4.38.1
modelscope~=1.12.0
transformers~=4.38.1
pypdf~=4.0.2
python-pptx~=0.6.23
flask~=3.0.0
nebula3-python==3.5.0
llama-index==0.10.10
scikit-learn==1.5.0
# cupy==13.3.0
llama-index==0.10.10
llama-index-vector-stores-milvus==0.1.3
llama-index-embeddings-huggingface==0.1.4
llama-index-llms-ollama==0.1.2
prettytable==3.11.0
socksio==1.0.0
overrides==7.7.0
cupy-cuda11x